+++ src/app/app-routing.module.ts
@@ -25,19 +25,12 @@ import {TableDemoComponent} from './demo/view/tabledemo.component';
 import {ListDemoComponent} from './demo/view/listdemo.component';
 import {AppTimelineDemoComponent} from './pages/app.timelinedemo.component';
 import {TreeDemoComponent} from './demo/view/treedemo.component';
-import {DisplayComponent} from './utilities/display.component';
-import {ElevationComponent} from './utilities/elevation.component';
-import {FlexboxComponent} from './utilities/flexbox.component';
-import {GridComponent} from './utilities/grid.component';
 import {IconsComponent} from './utilities/icons.component';
-import {WidgetsComponent} from './utilities/widgets.component';
-import {SpacingComponent} from './utilities/spacing.component';
-import {TypographyComponent} from './utilities/typography.component';
-import {TextComponent} from './utilities/text.component';
 import {AppCrudComponent} from './pages/app.crud.component';
 import {AppCalendarComponent} from './pages/app.calendar.component';
 import {AppInvoiceComponent} from './pages/app.invoice.component';
 import {AppHelpComponent} from './pages/app.help.component';
+import {BlocksComponent} from './blocks/blocks/blocks.component';
 
 @NgModule({
     imports: [
@@ -62,22 +55,15 @@ import {AppHelpComponent} from './pages/app.help.component';
                     {path: 'uikit/misc', component: MiscDemoComponent},
                     {path: 'uikit/charts', component: ChartsDemoComponent},
                     {path: 'uikit/file', component: FileDemoComponent},
-                    {path: 'utilities/display', component: DisplayComponent},
-                    {path: 'utilities/elevation', component: ElevationComponent},
-                    {path: 'utilities/flexbox', component: FlexboxComponent},
-                    {path: 'utilities/grid', component: GridComponent},
                     {path: 'utilities/icons', component: IconsComponent},
-                    {path: 'utilities/widgets', component: WidgetsComponent},
-                    {path: 'utilities/spacing', component: SpacingComponent},
-                    {path: 'utilities/typography', component: TypographyComponent},
-                    {path: 'utilities/text', component: TextComponent},
                     {path: 'pages/crud', component: AppCrudComponent},
                     {path: 'pages/calendar', component: AppCalendarComponent},
                     {path: 'pages/timeline', component: AppTimelineDemoComponent},
                     {path: 'pages/invoice', component: AppInvoiceComponent},
                     {path: 'pages/help', component: AppHelpComponent},
                     {path: 'pages/empty', component: EmptyDemoComponent},
-                    {path: 'documentation', component: DocumentationComponent}
+                    {path: 'documentation', component: DocumentationComponent},
+                    {path: 'blocks', component: BlocksComponent}
                 ]
             },
             {path: 'error', component: AppErrorComponent},